requirements.txt
google-adk
google-cloud-aiplatform[adk,agent_engines]
requests
python-dotenv
httpx
mcp
Pillow
